authorization

+++ app/Http/Controllers/Comment/CommentController.php
@@ -5,14 +5,21 @@ namespace App\Http\Controllers\Comment;
 use App\Comment;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function create()
     {
         $comment = new Comment;
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -21,8 +28,13 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back();
+        if (Gate::allows('comment-delete', $comment)) {
+            $comment->delete();
+            return back();
+        } else {
+            return back()->with('error', 'Unauthorized');
+        }
+        
     }
 }
